added script and config folders to xc_pipeline, updated calls to match new structure.
