+++ xc_configs/config_curruca_communis.yaml
