+++ xc_configs/config_parus_major.yaml
