+++ xc_configs/config_phylloscopus_collybita.yaml
