+++ xc_configs/config_strix_aluco.yaml
